backend/database/requirements.txt
# Web Framework
fastapi==0.104.1         # Web framework - lets us create API endpoints
uvicorn==0.24.0          # Server - runs our FastAPI app

# Database - ORM (Object-Relational Mapping)
sqlalchemy==2.0.23       # ORM - work with database using Python objects
psycopg2-binary==2.9.9   # PostgreSQL driver - SQLAlchemy needs this

# Security - for authentication
bcrypt==4.1.1            # Password hashing
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4   # Password utilities

# Validation
pydantic[email]==2.5.0   # Data validation
python-multipart==0.0.6  # Form data

# Environment
python-dotenv==1.0.0     # Reads .env file - keeps secrets safe
